refactor: modules, naming
